+++ cmake/dispatchers.cmake
@@ -1,0 +1,5 @@
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/dispatchers/std.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/dispatchers/glib.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/dispatchers/glibmm.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/dispatchers/qt.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/dispatchers/freertos.cmake)