+++ cmake/install.cmake
@@ -1,0 +1,4 @@
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/install/glib.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/install/glibmm.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/install/std.cmake)
+include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/install/qt.cmake)